requirements.txt
absl-py==2.1.0
addict==2.4.0
appdirs==1.4.4
apptools==5.2.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
ConfigArgParse==1.7
configobj==5.0.8
cycler==0.12.1
Cython==3.0.9
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
docker-pycreds==0.4.0
easydict==1.9
envisage==7.0.3
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.15.4
Flask==3.0.2
fonttools==4.50.0
freetype-py==2.4.0
fsspec==2024.6.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.2
google-auth-oauthlib==1.0.0
grpcio==1.62.1
h5py==2.10.0
hydra-core==0.11.3
idna==3.6
imageio==2.34.0
importlib_metadata==7.0.2
importlib_resources==6.3.1
ipython==8.12.3
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lmdb==1.4.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mayavi==4.8.1
mpmath==1.3.0
msgpack==1.0.8
msgpack-numpy==0.4.8
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
numpy==1.23.1
oauthlib==3.2.2
omegaconf==1.4.1
open3d==0.18.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
plotly==5.20.0
# pointnet2-ops==3.0.0 using original repository
prompt-toolkit==3.0.43
protobuf==3.20.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyface==8.0.0
pyglet==2.0.14
Pygments==2.17.2
PyOpenGL==3.1.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyquaternion==0.9.9
pyrender==0.1.39
python-dateutil==2.9.0.post0
python-fcl==0.7.0.6
pytorch-lightning==0.7.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.4.0
retrying==1.3.4
rpds-py==0.18.0
rsa==4.9
Rtree==1.2.0
scikit-learn==1.3.2
scipy==1.8.1
sentry-sdk==1.43.0
setproctitle==1.3.3
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12.1
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.0
<<<<<<< HEAD
threadpoolctl
#torch==1.10.0+cu111
torchaudio
torchvision
tqdm
=======
threadpoolctl==3.4.0
tqdm==4.45.0
>>>>>>> 13173cbf73c031797020055a373b6adba893e6d8
traitlets==5.14.2
traits==6.4.3
traitsui==8.0.0
trimesh==4.2.0
triton==2.3.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vtk==9.3.0
wandb==0.15.11
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.41.2
widgetsnbextension==4.0.10
zipp==3.18.1
